Add better resolution for finding includes.
The issue is documented in the code comment. Essentially 'repo' does not
enforce that each manifest snippet is reasonable on its own by requiring
that any 'remote' element is defined in the manifest in which it is
referenced. To deal with this I've created a utility which makes it
easy to get all of the remotes defined across all the manifests through
their includes.

This also adds the concept of a 'parent' manifest so we can traverse the
manifest inclusion tree.